Added the four relevant files for this paper
NAIpaper1406.pdf - The first four tables of this paper are what I am trying to replicate. It has been highlighted and commented on for convenience.
NAI_DataAppendix.pdf - This is the appendix for the NAIpaper1406.pdf. It may contain useful information, though I haven't found much use of it yet.
OLS.dta - This is the source data used for all regressions and other work in this paper. No changes have been made to this file. Theoretically all tables should be replicable with the information in this file, but due to my lack of knowledge with programming, I had to manually create some of my own columns and variables.
Indian2000Data.csv - This is my manipulated version of the source data. I've added some columns here and there, in addition to (obviously) changing the file format.
IndianOLSData.xlsx - This file contains the Table 1 Replication, along with lots of weird data manipulation. Probably the least useful file, but included to prove that only using the Year 20